// availabili.tf/src/client/models/EventWithPlayerSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EventSchema } from './EventSchema';
import type { PlayerEventRolesSchema } from './PlayerEventRolesSchema';
export type EventWithPlayerSchema = {
    event: EventSchema;
    playerEvent: (PlayerEventRolesSchema | null);
};